First version of the in-progress admin panel
It looks up questions that have to be admitted, shows the 5 first questions, with input boxes to change the question or answers if needed, with a button accept the question or delete the question. The css will need to be improved and we need to do the sql queries for the aformentioned tasks.